+++ lib/edfize/signal.rb
@@ -1,7 +1,7 @@
 module Edfize
   class Signal
     attr_accessor :label, :transducer_type, :physical_dimension, :physical_minimum, :physical_maximum,
-                  :digital_minimum, :digital_maximum, :prefiltering, :samples_in_data_record, :reserved_area
+                  :digital_minimum, :digital_maximum, :prefiltering, :samples_in_data_record, :reserved_area, :samples
 
     def initialize
       @label = ''
@@ -14,6 +14,7 @@ module Edfize
       @prefiltering = ''
       @samples_in_data_record = ''
       @reserved_area = ''
+      @samples = []
     end
 
   end